pyproject.toml
[project]
name = "aipm"
version = "0.1.0"
description = "The AI Project Manager - manage large projects distributed over multiple issue trackers and tools"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "click>=8.1",
    "jira>=3.8",
    "PyGithub>=2.1",
    "github-copilot-sdk>=0.1",
    "toml>=0.10",
    "rich>=13.0",
    "gitpython>=3.1",
]

[project.scripts]
aipm = "aipm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["aipm"]

[tool.ty]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "ty>=0.0",
]
